% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation}
\alias{aggregation}
\title{function used to generate valid choices for \code{aggregation_function}}
\usage{
aggregation(values = c("mean", "min", "max"))
}
\arguments{
\item{values}{a few common choices for aggregation}
}
\description{
function used to generate valid choices for \code{aggregation_function}
}
