% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.extendedIsolationForest.R
\name{step_outliers_h2o.extendedIsolationForest}
\alias{step_outliers_h2o.extendedIsolationForest}
\alias{tidy.step_outliers_h2o.extendedIsolationForest}
\title{Calculate the \link[=h2o.extendedIsolationForest]{h2o package extendedIsolationForest} outlier "score"}
\usage{
step_outliers_h2o.extendedIsolationForest(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  outlier_score = NULL,
  columns = NULL,
  name_mutate = ".outliers_h2o.extendedIsolationForest",
  options = list(extension_level = "max"),
  init_options = list(),
  skip = TRUE,
  id = rand_id("outliers_h2o.extendedIsolationForest")
)

\method{tidy}{step_outliers_h2o.extendedIsolationForest}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{selections()}} for more details.}

\item{role}{not defined for this function}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{outlier_score}{a placeholder for the exit of this function don't change}

\item{columns}{A character string of variable names that will
be populated (eventually) by the terms argument.}

\item{name_mutate}{the name of the generated column with h2o.extendedIsolationForest scores}

\item{options}{a list with arguments to  \link[h2o:h2o.extendedIsolationForest]{h2o::h2o.extendedIsolationForest} function.}

\item{init_options}{a list with parameters to \link[h2o:h2o.init]{h2o::h2o.init}}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_outliers_h2o.extendedIsolationForest} object.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any), with the name on \code{name_mutate} and the scores calculated. For the
\code{tidy} method, a tibble with columns \code{index} (the row indexes of the tibble) and \code{outlier_score} (the scores).
}
\description{
\code{step_outliers_h2o.extendedIsolationForest} creates a \emph{specification} of a recipe
step that will calculate the outlier score using \link{h2o.extendedIsolationForest} from \code{h2o}.
}
\details{
All columns in the data are sampled and returned by \code{\link[=juice]{juice()}}
and \code{\link[=bake]{bake()}}.

When used in modeling, users should strongly consider using the
option \code{skip = TRUE} so that this operation is \emph{not}
conducted outside of the training set.
}
\examples{
library(recipes)
library(tidy.outliers)
rec_obj <-
  recipe(mpg ~ ., data = mtcars) \%>\%
  step_outliers_h2o.extendedIsolationForest(all_numeric(), -all_outcomes()) \%>\%
  prep(mtcars)

juice(rec_obj)

tidy(rec_obj, number = 1)
}
\concept{preprocessing}
\keyword{datagen}
