% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maha.R
\name{step_outliers_maha}
\alias{step_outliers_maha}
\alias{tidy.step_outliers_maha}
\title{Calculate the \link[=maha]{Mahalanobis} outlier "probability"}
\usage{
step_outliers_maha(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  outlier_probability = NULL,
  name_mutate = ".outliers_maha",
  options = list(cutoff = 0, rnames = FALSE),
  skip = TRUE,
  id = rand_id("outliers_maha")
)

\method{tidy}{step_outliers_maha}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables will be transformed. See \code{\link[=selections]{selections()}} for
more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{not defined for this function}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{outlier_probability}{a placeholder for the exit of this function don't change}

\item{name_mutate}{the name of the generated column with maha probabilities}

\item{options}{a list with cutoff point which is ignored and the rnames argument of the \link{maha} function}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake.recipe()}}? While all operations are baked
when \code{\link[recipes:prep]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_outliers_maha} object.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any), with the name on \code{name_mutate} and the probabilities calculated. For the
\code{tidy} method, a tibble with columns \code{index} (the row indexes of the tibble) and \code{outlier_probability} (the probabilites).
}
\description{
\code{step_outliers_maha} creates a \emph{specification} of a recipe
step that will calculate the probability of the row of selected columns being an outliers using \link{maha} from \code{OutlierDetection}.
}
\details{
All columns in the data are sampled and returned by \code{\link[=juice]{juice()}}
and \code{\link[=bake]{bake()}}.

All columns used in this step must be numeric with no missing data.

When used in modeling, users should strongly consider using the
option \code{skip = TRUE} so that this operation is \emph{not}
conducted outside of the training set.
}
\examples{
library(recipes)
library(tidy.outliers)
library(OutlierDetection)
rec_obj <-
  recipe(mpg ~ ., data = mtcars) \%>\%
  step_outliers_maha(all_numeric(), -all_outcomes()) \%>\%
  prep(mtcars)

juice(rec_obj)

tidy(rec_obj, number = 1)
}
\concept{preprocessing}
\keyword{datagen}
