% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.outliers-package.R
\docType{package}
\name{tidy.outliers-package}
\alias{tidy.outliers}
\alias{tidy.outliers-package}
\title{tidy.outliers: What the Package Does (One Line, Title Case)}
\description{
Creates two tipes of tidy steps for outliers: "probability" caluculation and outlier removal, the idea is that you can combine multiple methods and use them easily on a workflows modeling case.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunocarlin/tidy.outliers}
  \item Report bugs at \url{https://github.com/brunocarlin/tidy.outliers/issues}
}

}
\author{
\strong{Maintainer}: Bruno Carlin \email{brunotestaguzzat@gmail.com} (\href{https://orcid.org/0000-0002-7701-2953}{ORCID})

}
\keyword{internal}
